+++ bin/src/templates.dart
@@ -15,15 +15,15 @@ extension SourceAbstractControlExtension<C extends AbstractControl<T>, T> on C {
 final class AbstractControlSource<C extends AbstractControl<T>, T> {
   final C _control;
 
-  SourceListenable<ControlStatus> get status => _onStatus.select(_status);
-  SourceListenable<bool> get pristine => _onStatus.select(_pristine);
-  SourceListenable<bool> get dirty => _onStatus.select(_dirty);
-  SourceListenable<Map<String, Object>> get errors => _onStatus.select(_errors);
-  SourceListenable<bool> get hasErrors => _onStatus.select(_hasErrors);
+  SourceListenable<ControlStatus> get status => _onStatus(null, _status);
+  SourceListenable<bool> get pristine => _onStatus(null, _pristine);
+  SourceListenable<bool> get dirty => _onStatus(null, _dirty);
+  SourceListenable<Map<String, Object>> get errors => _onStatus(null, _errors);
+  SourceListenable<bool> get hasErrors => _onStatus(null, _hasErrors);
   SourceListenable<bool> hasError(String errorCode, [String? path]) =>
-      _onStatus.selectWith((errorCode, path), _hasError);
+      _onStatus((errorCode, path), _hasError);
   SourceListenable<Object?> getError(String errorCode, [String? path]) =>
-      _onStatus.selectWith((errorCode, path), _getError);
+      _onStatus((errorCode, path), _getError);
 
   SourceListenable<T?> get value => _AbstractControlValueSourceListenable(_control);
   SourceListenable<bool> get isEmpty => value.select(_isEmpty);
@@ -32,14 +32,14 @@ final class AbstractControlSource<C extends AbstractControl<T>, T> {
 
   AbstractControlSource._(this._control);
 
-  SourceListenable<AbstractControl<Object?>> get _onStatus =>
-      _AbstractControlStatusSourceListenable(_control);
+  SourceListenable<R> _onStatus<A, R>(A arg, R Function(A arg, C control) selector) =>
+      _AbstractControlStatusSourceListenable(_control, arg, selector);
 
-  static ControlStatus _status(AbstractControl<Object?> control) => control.status;
-  static bool _pristine(AbstractControl<Object?> control) => control.pristine;
-  static bool _dirty(AbstractControl<Object?> control) => control.dirty;
-  static Map<String, Object> _errors(AbstractControl<Object?> control) => control.errors;
-  static bool _hasErrors(AbstractControl<Object?> control) => control.hasErrors;
+  static ControlStatus _status(_, AbstractControl<Object?> control) => control.status;
+  static bool _pristine(_, AbstractControl<Object?> control) => control.pristine;
+  static bool _dirty(_, AbstractControl<Object?> control) => control.dirty;
+  static Map<String, Object> _errors(_, AbstractControl<Object?> control) => control.errors.cast();
+  static bool _hasErrors(_, AbstractControl<Object?> control) => control.hasErrors;
   static bool _hasError((String, String?) arg, AbstractControl<Object?> control) =>
       control.hasError(arg.$1, arg.$2);
   static Object? _getError((String, String?) arg, AbstractControl<Object?> control) =>
@@ -82,31 +82,37 @@ extension ControlStatusSourceExtensions on SourceListenable<ControlStatus> {
   static bool _enabled(ControlStatus status) => !_disabled(status);
 }
 
-final class _AbstractControlStatusSourceListenable
-    extends SourceListenable<AbstractControl<Object?>> {
-  final AbstractControl<Object?> _control;
+final class _AbstractControlStatusSourceListenable<C extends AbstractControl<Object?>, A, R>
+    extends SourceListenable<R> {
+  final C _control;
+  final A _arg;
+  final R Function(A arg, C control) _selector;
 
-  _AbstractControlStatusSourceListenable(this._control);
+  _AbstractControlStatusSourceListenable(this._control, this._arg, this._selector);
 
   @override
-  SourceSubscription<AbstractControl<Object?>> listen(
-    SourceListener<AbstractControl<Object?>> listener,
-  ) {
+  SourceSubscription<R> listen(SourceListener<R> listener) {
+    var current = _selector(_arg, _control);
     final subscription = _control.statusChanged.listen((_) {
-      listener(_control, _control);
+      final previous = current;
+      current = _selector(_arg, _control);
+      if (previous == current) return;
+      listener(previous, current);
     });
-    return SourceSubscriptionBuilder(() => _control, subscription.cancel);
+    return SourceSubscriptionBuilder(() => current, subscription.cancel);
   }
 
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is _AbstractControlStatusSourceListenable &&
+      other is _AbstractControlStatusSourceListenable<C, A, R> &&
           runtimeType == other.runtimeType &&
-          identical(_control, other._control);
+          identical(_control, other._control) &&
+          _arg == other._arg &&
+          identical(_selector, other._selector);
 
   @override
-  int get hashCode => Object.hash(runtimeType, _control);
+  int get hashCode => Object.hash(runtimeType, _control, _arg, _selector);
 }
 
 final class _AbstractControlValueSourceListenable<T> extends SourceListenable<T?> {