+++ test/sources/listenable_source_test.dart
@@ -27,6 +27,7 @@ void main() {
 
       expect(states, [1, 2]);
       subscription.cancel();
+      subscription.cancel(); // Ensure can cancel subscription many times
     });
 
     test('should not emit a value if the value is the same', () {
@@ -41,6 +42,7 @@ void main() {
 
       expect(states, isEmpty);
       subscription.cancel();
+      subscription.cancel(); // Ensure can cancel subscription many times
     });
 
     test('subscription.read() should return the current value', () {
@@ -54,6 +56,7 @@ void main() {
 
       expect(subscription.read(), 5);
       subscription.cancel();
+      subscription.cancel(); // Ensure can cancel subscription many times
     });
 
     test('should stop listening when the subscription is cancelled', () {
@@ -65,6 +68,7 @@ void main() {
 
       notifier.value = 1;
       subscription.cancel();
+      subscription.cancel(); // Ensure can cancel subscription many times
       notifier.value = 2;
 
       expect(states, [1]);
@@ -94,6 +98,7 @@ void main() {
       expect(states, [1]);
 
       subscription.cancel();
+      subscription.cancel(); // Ensure can cancel subscription many times
     });
   });